requirements_base.txt
# undetected-chromedriver
# nuitka @ https://github.com/Nuitka/Nuitka/archive/develop.zip
# botasaurus
# selenium-wire
# chromedriver_autoinstaller_fix
selenium-stealth
webdriver-manager
# selenium-authenticated-proxy
datetime
pytz
babel
colorama
pynput
SQLAlchemy
beautifulsoup4
requests
tzlocal
urllib3
certifi
maxminddb
geoip2
python-dateutil
joblib
numpy
socks
pyyaml
# proxy-checker @ git+https://github.com/dimaslanjaka/proxy-checker-python@master
-e file:packages/proxy-checker-python
-e file:packages/proxy-hunter-python
autopep8
pycodestyle
black
zstandard
selenium
setuptools
pycurl
rarfile
pycryptodome
django
django-filter
django-environ
djangorestframework
djangorestframework_simplejwt
django-cors-middleware
django-cors-headers
django-extensions
PyJWT
six
channels
python-dotenv
